Fix CI: oxfmt, knip, spellcheck, typecheck
Remove checkWithOxc.js (new-JS policy), consolidate in .mjs, format
benchmark JSON, add cspell ignore for sysctl keys, and register .mjs
scripts as knip entry points.
